Add codegen for mir statements and entities

+++ spade-mir/src/verilog.rs
@@ -2,7 +2,7 @@ pub fn assign(target: &str, value: &str) -> String {
     format!("assign {} = {};", target, value)
 }
 
-pub fn size_spec(size: u128) -> String {
+pub fn size_spec(size: u64) -> String {
     if size == 1 {
         format!("")
     } else {
@@ -10,9 +10,9 @@ pub fn size_spec(size: u128) -> String {
     }
 }
 
-pub fn wire(name: &str, size: u128) -> String {
+pub fn wire(name: &str, size: u64) -> String {
     format!("wire{} {};", size_spec(size), name)
 }
-pub fn reg(name: &str, size: u128) -> String {
+pub fn reg(name: &str, size: u64) -> String {
     format!("reg{} {};", size_spec(size), name)
 }
